requirements.txt
streamlit==1.55.0
tensorflow==2.15.0
numpy==1.25.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0